Header, Nav Bar and CSS
On my home page I have added the page header and the navigation bar. I have got my header to the size I want it so it fits the page properly. I have got teh template of the nav bar there but I am still refining it to make it look how I want it to and so that it matches the style of the rest of the website